Add test to new components
