+++ react-time-track/src/index.js
@@ -3,14 +3,12 @@ import React from "react";
 import { render } from "react-dom";
 import { Router, Redirect } from "@reach/router";
 import { jsx, Global } from "@emotion/core";
-import { FaBars } from "react-icons/fa";
-
 import { UserProvider } from "./contexts/user";
+
 import Login from "./views/login";
+import Home from "./views/home";
 import AllProjects from "./views/all-projects";
 import History from "./views/history";
-import NavBar from "./components/navbar";
-import Header from "./components/header";
 import UserList from "./components/user-list";
 
 const global = {
@@ -31,90 +29,42 @@ const global = {
   }
 };
 
-const gridContainer = {
-  display: "grid",
-  gridTemplateColumns: "1fr",
-  gridTemplateRows: "50px 1fr",
-  gridTemplateAreas: '"header" "main"',
-  height: "100vh",
-  "@media (min-width: 960px)": {
-    gridTemplateColumns: "240px 1fr",
-    gridTemplateAreas: '"navbar header" "navbar main"'
-  }
-};
-
-const menuIcon = {
-  position: "fixed",
-  display: "flex",
-  top: "5px",
-  left: "10px",
-  alignItems: "center",
-  justifyContent: "center",
-  backgroundColor: "#DADAE3",
-  borderRadius: "50%",
-  zIndex: "1",
-  cursor: "pointer",
-  padding: "12px"
-};
-
 function App() {
   const [currentUser, setCurrentUser] = React.useState(
     JSON.parse(localStorage.getItem("user")) || {}
   );
-
-  const [navBarActive, setNavBarActive] = React.useState("translateX(-245px)");
 
   function handleCurrentUser(newValue) {
     localStorage.setItem("user", JSON.stringify(newValue));
     setCurrentUser(newValue);
   }
 
-  function togleNavBar() {
-    navBarActive === "translateX(-245px)"
-      ? setNavBarActive("translateX(0)")
-      : setNavBarActive("translateX(-245px)");
-  }
   return (
     <UserProvider user={currentUser} setUser={handleCurrentUser}>
       <Global styles={global} />
       <Router>
+        {currentUser.name ? (
+          <Redirect
+            from="/login"
+            to={
+              window.location.pathname === "/login"
+                ? "/"
+                : window.location.pathname
+            }
+            noThrow
+          />
+        ) : (
+          window.location.pathname !== "/login" && (
+            <Redirect from={window.location.pathname} to="/login" noThrow />
+          )
+        )}
         <Login path="/login" />
+        <Home path="/">
+          <AllProjects path="/" />
+          <History path="/history" />
+          <UserList path="/members" />
+        </Home>
       </Router>
-      {currentUser.name ? (
-        <>
-          <Redirect from="/login" to="/" noThrow />
-          <div css={gridContainer}>
-            <div css={menuIcon} onClick={togleNavBar}>
-              <FaBars />
-            </div>
-            <Router
-              css={{
-                gridArea: "header"
-              }}
-            >
-              <Header tittle="Projects" path="/" />
-              <Header tittle="History" path="/history" />
-              <Header tittle="Members" path="/members" />
-            </Router>
-
-            <NavBar navBarActive={navBarActive} togleNavBar={togleNavBar} />
-            <Router
-              css={{
-                gridArea: "main",
-                marginTop: "0.5em"
-              }}
-            >
-              <AllProjects path="/" />
-              <History path="/history" />
-              <UserList path="/members" />
-            </Router>
-          </div>
-        </>
-      ) : (
-        window.location.pathname !== "/login" && (
-          <Redirect from={window.location.pathname} to="/login" noThrow />
-        )
-      )}
     </UserProvider>
   );
 }
